+++ cookbooks/aws-parallelcluster-config/recipes/head_node_base.rb
@@ -15,10 +15,17 @@
 # OR CONDITIONS OF ANY KIND, express or implied. See the License for the specific language governing permissions and
 # limitations under the License.
 
-manage_ebs "ebs" do
+# generate the shared storages mapping file
+template node['cluster']['shared_storages_mapping_path'] do
+  source 'shared_storages/shared_storages_data.erb'
+  mode '0644'
+end
+
+manage_ebs "add ebs" do
   shared_dir_array node['cluster']['ebs_shared_dirs'].split(',')
   vol_array node['cluster']['volume'].split(',')
   action %i(mount export)
+  not_if { node['cluster']['ebs_shared_dirs'].split(',').empty? }
 end
 
 # Export /home