+++ third-party/selinux-6.0.5/test/cookbooks/selinux_test/recipes/install.rb
@@ -1,0 +1,3 @@
+apt_update 'default' if platform_family?('debian')
+
+selinux_install 'install packages'